Added files via upload
So this is the product backlog so far. We don't really have any sprints just yet, they all have the same priority for now so that's also void for the time being, and we can't make out the size right now until we do the actual work, but this should be fine right?