Create Game.java
There are missing methods within other classes that will need to be implemented for this code to run correctly. A .getDicePairs() method is needed and it will need to return a DicePairs array. Also a .getDice method may be needed in order to use the .rollDice() method in Turn.

Comments will be added to help with clarity of the code's functionality for future versions.