Added files via upload
Hey guys, the gui I've been working on so far is finally here! I've still got quite a bit of work to do with this big guy, but the code I'm uploading to you now has good implementation for the beginning of the game right before you start rolling dice. I'm currently working of the pairings and the attacks, then I'll try to work on the endgame, but for now I really want to figure out how to differentiate panels for each player. What I mean by that is when the players are all choosing their pairs in turns, I want player one's dice rolled to show during their decision to pair, and then player two for their turn, and so on. And then I want their dice pairs to show up for their attacks as well. So essentially I want to reuse the panels (and I'm pretty sure I know how to do that now in terms of resetting the panel instead of looping it) but I also want the dice pertaining to their respective player to show up for their panels. I'm also thinking the name for the game should be Dicecapades. Sor