Added files via upload
Made a series of modifications to distinguish computer players from those who are human controlled and how they roll. Also enable rerolling.
Some of these changes are tentative and will be revised as we hook up these things to the GUI but for now I'm using the console for when input is needed.